Export ghvs as module
